requirements.txt
fastapi==0.88.0
uvicorn==0.20.0

PyJWT==2.6.0
python-decouple==3.6

email-validator==1.3.0